Move more model into models/user
