+++ models/userlist_test.go
@@ -64,32 +64,5 @@ func testUserListIsUserOrgOwner(t *testing.T, orgID int64, expected map[int64]bo
 	assert.NoError(t, err)
 	members, _, err := org.GetMembers()
 	assert.NoError(t, err)
-	assert.Equal(t, expected, members.IsUserOrgOwner(orgID))
-}
-
-func TestUserListIsTwoFaEnrolled(t *testing.T) {
-	assert.NoError(t, unittest.PrepareTestDatabase())
-	tt := []struct {
-		orgid    int64
-		expected map[int64]bool
-	}{
-		{3, map[int64]bool{2: false, 4: false, 28: false}},
-		{6, map[int64]bool{5: false, 28: false}},
-		{7, map[int64]bool{5: false}},
-		{25, map[int64]bool{24: true}},
-		{22, map[int64]bool{}},
-	}
-	for _, v := range tt {
-		t.Run(fmt.Sprintf("IsTwoFaEnrolledOfOrdIg%d", v.orgid), func(t *testing.T) {
-			testUserListIsTwoFaEnrolled(t, v.orgid, v.expected)
-		})
-	}
-}
-
-func testUserListIsTwoFaEnrolled(t *testing.T, orgID int64, expected map[int64]bool) {
-	org, err := GetOrgByID(orgID)
-	assert.NoError(t, err)
-	members, _, err := org.GetMembers()
-	assert.NoError(t, err)
-	assert.Equal(t, expected, members.GetTwoFaStatus())
+	assert.Equal(t, expected, IsUserOrgOwner(members, orgID))
 }
